Move service configuration code to `config` package
As these are all related to the configuration of the service, we should
treat them as such and put them into the `config` package.
